back-end/hogword/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
supabase==2.9.0
python-dotenv==1.0.1
pydantic==2.6.1
email-validator==2.1.0.post1
